+++ CHANGELOG.md
@@ -4,9 +4,9 @@
 
 - Use bract.core 0.5.1
   - [TODO] Use the GA version
-- [TODO] Fix commands visibility in CLI help message
+- Fix commands visibility in CLI help message
   - In root CLI inducers move `read-context` ahead of `parse-args`
-  - [TODO] Add `:bract.cli/pre-inducers` to run before parsing CLI args
+  - Run inducers `:bract.cli/pre-inducers` before parsing CLI args
 - Make `bract.cli.main/-main` return exit code using `:bract.core/exit-code`